+++ src/index.js
@@ -2,7 +2,10 @@ import * as utils from './utils.js';
 
 function getDefaults() {
   return {
-    debounceInterval: 50
+    debounceInterval: 50,
+
+    // Default: true; (Backwards compatible). This option allows to fetch namespaces for multiple languages at once or to fetch them per language
+    multiLanguage: true
   };
 }
 
@@ -24,7 +27,8 @@ class Backend {
       this.backend = this.backend || utils.createClassOnDemand(this.options.backend);
       this.backend.init(services, this.options.backendOption, i18nextOptions);
     }
-    if (this.backend && !this.backend.readMulti) throw new Error('The wrapped backend does not support the readMulti function.')
+    if (this.backend && !this.backend.readMulti)
+      throw new Error('The wrapped backend does not support the readMulti function.');
   }
 
   read(language, namespace, callback) {
@@ -40,26 +44,72 @@ class Backend {
   load() {
     if (!this.backend || !this.pending.length) return;
 
-    // reset pending
-    const loading = this.pending;
+    // Reset pending
+    const loading = this.pending; // [{ language, namespace, callback }]
     this.pending = [];
 
-    // get all languages and namespaces needed to be loaded
-    const toLoad = loading.reduce((mem, item) => {
-      if (mem.languages.indexOf(item.language) < 0) mem.languages.push(item.language);
-      if (mem.namespaces.indexOf(item.namespace) < 0) mem.namespaces.push(item.namespace);
-      return mem;
-    }, { languages: [], namespaces: [] })
+    // Build the lang/namespace/callback object
+    // { language: { namespace: callback } }
+    const namespacePerLanguage = {};
+    loading.forEach(({ language, namespace, callback }) => {
+      namespacePerLanguage[language] = namespacePerLanguage[language] || {};
+      namespacePerLanguage[language][namespace] =
+        namespacePerLanguage[language][namespace] || callback; // Just making sure we do not have duplicates
+    });
+
+    let toLoad; // [{ languages, namespaces }]
+    // Create only 1 object as we fetch everything at once
+    if (this.options.multiLanguage) {
+      toLoad = [
+        {
+          languages: Object.keys(namespacePerLanguage),
+          namespaces: Object.keys(namespacePerLanguage).reduce((acc, language) => {
+            const namespaces = Object.keys(namespacePerLanguage[language]);
+
+            namespaces.forEach(ns => {
+              if (acc.indexOf(ns) === -1) {
+                acc.push(ns);
+              }
+            });
+
+            return acc;
+          }, [])
+        }
+      ];
+      // Create 1 object per language as we fetch multiple namespace per language
+    } else {
+      toLoad = Object.keys(namespacePerLanguage).map(language => {
+        return { languages: [language], namespaces: Object.keys(namespacePerLanguage[language]) };
+      });
+    }
 
     // load
-    this.backend.readMulti(toLoad.languages, toLoad.namespaces, (err, data) => {
-      if (err) return loading.forEach(item => item.callback(err, data));
+    toLoad.forEach(({ languages, namespaces }) => {
+      this.backend.readMulti(languages, namespaces, (err, data) => {
+        languages.forEach(language => {
+          namespaces.forEach(namespace => {
+            const callback = (namespacePerLanguage[language] || {})[namespace];
 
-      loading.forEach(item => {
-        const translations = data[item.language] && data[item.language][item.namespace];
-        item.callback(null, translations || {}); // if no error and no translations for that lng-ns pair return empty object
+            if (!callback) {
+              return null;
+            }
+
+            if (err) {
+              return callback(err, null);
+            }
+
+            let translations;
+            if (this.options.multiLanguage) {
+              translations = (data[language] || {})[namespace] || {};
+            } else {
+              translations = data[namespace] || {};
+            }
+
+            callback(null, translations);
+          });
+        });
       });
-    })
+    });
   }
 
   create(languages, namespace, key, fallbackValue) {
@@ -69,6 +119,5 @@ class Backend {
 }
 
 Backend.type = 'backend';
-
 
 export default Backend;