+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+### Upcoming
+
+- `multiLanguage` option
+
 ### 0.1.1
+
 - initial version